% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{similar_chars}
\alias{similar_chars}
\title{Calculate similar characters of two string vectors, element-by-element.}
\usage{
similar_chars(x, y)
}
\arguments{
\item{x}{A character vector.}

\item{y}{A character vector.}
}
\value{
A vector for the element-wise amount of similar characters between
\code{x} and \code{y}.
}
\description{
Calculate similar characters of two string vectors, element-by-element.
}
\examples{
similar_chars(c('dianhua', 'gandong'), c('denwa', 'kando'))
}

