% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{similar_chars_single}
\alias{similar_chars_single}
\title{Count similar characters of two strings (only for length one vectors).}
\usage{
similar_chars_single(x, y)
}
\arguments{
\item{x}{A character vector of length one.}

\item{y}{A character vector of length one.}
}
\value{
A length one vector for the amount of similar characters between
\code{x} and \code{y}.
}
\description{
Count similar characters of two strings (only for length one vectors).
}

