% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/func.R
\name{similarity}
\alias{similarity}
\title{Calculate phonological similarity of two string vectors, element-by-element.}
\usage{
similarity(x, y)
}
\arguments{
\item{x}{A character vector.}

\item{y}{A character vector.}
}
\value{
A vector for the element-wise phonological similarity between
\code{x} and \code{y}.
}
\description{
Calculate phonological similarity of two string vectors, element-by-element.
}
\examples{
similarity(c('dianhua', 'gandong'), c('denwa', 'kando'))
}

